Precision Calculation and Visualisations.
Added functions to calculate the precision of the model on this dataset with a precision to number of neighbors requested graph. Used PCA to reduce feature-length and visualized the clustering of the features with t-sne. The next steps can be to visualize differences with u-map or finetune the model used to extract features.